requirements.txt
httpx[http2]>=0.27
beautifulsoup4>=4.12
lxml>=5.0
markdownify>=0.12
PyYAML>=6.0
tqdm>=4.66
